feat: implement order details functionality (#43)

+++ Models/CartItemModel.cs
@@ -2,7 +2,7 @@
 {
     public class CartItemModel
     {
-        public long ProductId { get; set; }
+        public int ProductId { get; set; }
         public string ProductName { get; set; }
         public string ImageUrl { get; set; }
         public int Quantity { get; set; }
